refactor: Move mockOrganizationResponseForAffiliation to fixtures
