+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/service/CandidateNotesTest.java
@@ -1,8 +1,8 @@
 package no.sikt.nva.nvi.common.service;
 
 import static no.sikt.nva.nvi.common.RequestFixtures.createNoteRequest;
+import static no.sikt.nva.nvi.common.model.OrganizationFixtures.mockOrganizationResponseForAffiliation;
 import static no.sikt.nva.nvi.test.TestUtils.createUpsertCandidateRequest;
-import static no.sikt.nva.nvi.test.TestUtils.mockOrganizationResponseForAffiliation;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 import static no.unit.nva.testutils.RandomDataGenerator.randomUri;
 import static org.hamcrest.MatcherAssert.assertThat;
